test(core): split build plugin tests
